Safari-611.1.10/LayoutTests/fast/text/international/system-language/navigator-language/navigator-language-hi.html

<!DOCTYPE html><!-- webkit-test-runner [ language=hi ] -->
<html>
<head>
<script src="../../../../../resources/js-test.js"></script>
</head>
<body>
<script>
shouldBeEqualToString("navigator.language", "hi-in");
</script>
</body>
</html>
